requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
install==1.3.5
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.2
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langsmith==0.1.64
llvmlite==0.42.0
MarkupSafe==2.1.5
mlx==0.14.0; sys_platform == 'darwin' and platform_machine == 'arm64'
mlx-whisper==0.1.0; sys_platform == 'darwin' and platform_machine == 'arm64'
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scipy==1.13.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sympy==1.12.1
tenacity==8.3.0
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
